# openapi.json
{"openapi":"3.0.0","servers":[{"url":"http://example.com","description":"","variables":{}}],"info":{"version":"9acc2264-ec91-4ec4-aaf3-ab10361d8536","title":"posty","description":"","termsOfService":"","contact":{},"license":{"name":""}},"paths":{"/users":{"post":{"summary":"Create User","operationId":"CreateUser","parameters":[],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":{"username":"test"}},"example":"{\n    \"username\": \"test\" \n}"}}},"tags":["Users"]}},"/users/{id}":{"get":{"summary":"Get User","operationId":"GetUser","parameters":[{"$ref":"#/components/parameters/id"}],"responses":{"200":{"description":"","headers":{}}},"tags":["Users"]}},"/users/{id}/follow":{"post":{"summary":"Follow user","operationId":"Followuser","parameters":[{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"number"}}],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":{"follower_id":"6223c394df1ec5873cbe22a1"}},"example":"{\n    \"follower_id\": \"6223c394df1ec5873cbe22a1\"\n}"}}},"tags":["Users"]}},"/users/{id}/unfollow":{"post":{"summary":"Unfollow user","operationId":"Unfollowuser","parameters":[{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"number"}}],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":{"follower_id":"6223c394df1ec5873cbe22a1"}},"example":"{\n    \"follower_id\": \"6223c394df1ec5873cbe22a1\"\n}"}}},"tags":["Users"]}},"/users/{id}/posts":{"post":{"summary":"Create Post","operationId":"CreatePost","parameters":[{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"string","example":"6224cac190e3a3d9861caeaa"}}],"responses":{"200":{"description":"","headers":{}}},"requestBody":{"required":true,"content":{"text/plain":{"schema":{"type":"string","example":{"content":"this","parent_id":"6225301efed428f410c97689"}},"example":"{\n    \"content\" : \"this\",\n    \"parent_id\": \"6225301efed428f410c97689\"\n}"}}},"tags":["Posts"]},"get":{"summary":"Get User's last Posts","operationId":"GetUser'slastPosts","parameters":[{"$ref":"#/components/parameters/page"},{"$ref":"#/components/parameters/limit"},{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"string","example":"6224cac190e3a3d9861caeaa"}}],"responses":{"200":{"description":"","headers":{}}},"tags":["Posts"]}},"/users/{id}/feed":{"get":{"summary":"Get user's feed","operationId":"Getuser'sfeed","parameters":[{"name":"page","in":"query","required":true,"style":"form","schema":{"type":"number"}},{"name":"limit","in":"query","required":true,"style":"form","schema":{"type":"number"}},{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"string","example":"6226a3530ddc894a187d1860"}}],"responses":{"200":{"description":"","headers":{}}},"tags":["Posts"]}}},"components":{"parameters":{"id":{"name":"id","in":"path","required":true,"style":"simple","schema":{"type":"string","example":"6226528d608e2c92f235bd0a"}},"page":{"name":"page","in":"query","required":true,"style":"form","schema":{"type":"number"}},"limit":{"name":"limit","in":"query","required":true,"style":"form","schema":{"type":"number"}}}},"security":[],"tags":[],"externalDocs":{"url":"","description":""},"warnings":[]}